requirements.txt
streamlit
streamlit-webrtc

numpy
pandas
pillow
matplotlib

opencv-python
av

scikit-learn
joblib

tensorflow
keras

ultralytics
torch
torchvision
torchaudio
